Move register_(cmaps|cycles|...) to config.py
