+++ proplot/colors.py
@@ -35,7 +35,6 @@ __all__ = [
     'LinearSegmentedColormap',
     'ListedColormap',
     'MidpointNorm', 'PerceptuallyUniformColormap',
-    'cmaps', 'colors', 'cycles', 'fonts',
     'make_mapping_array',
     'Colormap', 'Colors', 'Cycle', 'Norm',
 ]
@@ -2918,267 +2917,9 @@ def _from_file(filename, listed=False, warn_on_failure=False):
     return cmap
 
 
-@docstring.add_snippets
-def register_cmaps(user=True, default=False):
-    """
-    Register colormaps packaged with ProPlot or saved to the
-    ``~/.proplot/cmaps`` folder. This is called on import.
-    Colormaps are registered according to their filenames -- for example,
-    ``name.xyz`` will be registered as ``'name'``.
-
-    %(register.ext_table)s
-
-    To visualize the registered colormaps, use `~proplot.show.show_cmaps`.
-
-    Parameters
-    ----------
-    %(register_cmaps.params)s
-    """
-    for i, dirname, filename in _iter_data_paths('cmaps', user=user, default=default):
-        path = os.path.join(dirname, filename)
-        cmap = pcolors.LinearSegmentedColormap.from_file(path, warn_on_failure=True)
-        if not cmap:
-            continue
-        if i == 0 and cmap.name.lower() in (
-            'phase', 'graycycle', 'romao', 'broco', 'corko', 'viko',
-        ):
-            cmap.set_cyclic(True)
-        pcolors._cmapdict[cmap.name] = cmap
 
 
-@docstring.add_snippets
-def register_cycles(user=True, default=False):
-    """
-    Register color cycles packaged with ProPlot or saved to the
-    ``~/.proplot/cycles`` folder. This is called on import. Color cycles
-    are registered according to their filenames -- for example, ``name.hex``
-    will be registered as ``'name'``.
 
-    %(register.ext_table)s
-
-    To visualize the registered color cycles, use `~proplot.show.show_cycles`.
-
-    Parameters
-    ----------
-    %(register_cycles.params)s
-    """
-    for _, dirname, filename in _iter_data_paths('cycles', user=user, default=default):
-        path = os.path.join(dirname, filename)
-        cmap = pcolors.ListedColormap.from_file(path, warn_on_failure=True)
-        if not cmap:
-            continue
-        pcolors._cmapdict[cmap.name] = cmap
-
-
-@docstring.add_snippets
-def register_colors(user=True, default=False, space='hcl', margin=0.10):
-    """
-    Register the `open-color <https://yeun.github.io/open-color/>`_ colors,
-    XKCD `color survey <https://xkcd.com/color/rgb/>`_ colors, and colors
-    saved to the ``~/.proplot/colors`` folder. This is called on import.
-    The color survey colors are filtered to a subset that is "perceptually
-    distinct" in the HCL colorspace. The user color names are loaded from
-    ``.txt`` files saved in ``~/.proplot/colors``. Each file should contain
-    one line per color in the format ``name : hex``. Whitespace is ignored.
-
-    To visualize the registered colors, use `~proplot.show.show_colors`.
-
-    Parameters
-    ----------
-    %(register_colors.params)s
-    space : {'hcl', 'hsl', 'hpl'}, optional
-        The colorspace used to detect "perceptually distinct" colors.
-    margin : float, optional
-        The margin by which a color's normalized hue, saturation, and
-        luminance channel values must differ from the normalized channel
-        values of the other colors to be deemed "perceptually distinct."
-    """
-    # Reset native colors dictionary
-    mcolors.colorConverter.colors.clear()  # clean out!
-    mcolors.colorConverter.cache.clear()  # clean out!
-
-    # Add in base colors and CSS4 colors so user has no surprises
-    for name, dict_ in (('base', BASE_COLORS), ('css', mcolors.CSS4_COLORS)):
-        mcolors.colorConverter.colors.update(dict_)
-
-    # Load colors from file and get their HCL values
-    # NOTE: Colors that come *later* overwrite colors that come earlier.
-    hex = re.compile(rf'\A{pcolors.HEX_PATTERN}\Z')  # match each string
-    for i, dirname, filename in _iter_data_paths('colors', user=user, default=default):
-        path = os.path.join(dirname, filename)
-        cat, ext = os.path.splitext(filename)
-        if ext != '.txt':
-            raise ValueError(
-                f'Unknown color data file extension ({path!r}). '
-                'All files in this folder should have extension .txt.'
-            )
-
-        # Read data
-        loaded = {}
-        with open(path, 'r') as fh:
-            for cnt, line in enumerate(fh):
-                # Load colors from file
-                stripped = line.strip()
-                if not stripped or stripped[0] == '#':
-                    continue
-                pair = tuple(
-                    item.strip().lower() for item in line.split(':')
-                )
-                if len(pair) != 2 or not hex.match(pair[1]):
-                    warnings._warn_proplot(
-                        f'Illegal line #{cnt + 1} in file {path!r}:\n'
-                        f'{line!r}\n'
-                        f'Lines must be formatted as "name: hexcolor".'
-                    )
-                    continue
-                # Never overwrite "base" colors with xkcd colors.
-                # Only overwrite with user colors.
-                name, color = pair
-                if i == 0 and name in BASE_COLORS:
-                    continue
-                loaded[name] = color
-
-        # Add every user color and every opencolor color and ensure XKCD
-        # colors are "perceptually distinct".
-        if i == 1:
-            mcolors.colorConverter.colors.update(loaded)
-        elif cat == 'opencolor':
-            mcolors.colorConverter.colors.update(loaded)
-            OPEN_COLORS.update(loaded)
-        elif cat == 'xkcd':
-            # Always add these colors, but make sure not to add other
-            # colors too close to them.
-            hcls = []
-            filtered = []
-            for name in ALWAYS_ADD:
-                color = loaded.pop(name, None)
-                if color is None:
-                    continue
-                if 'grey' in name:
-                    name = name.replace('grey', 'gray')
-                hcls.append(to_xyz(color, space=space))
-                filtered.append((name, color))
-                mcolors.colorConverter.colors[name] = color
-                XKCD_COLORS[name] = color
-
-            # Get locations of "perceptually distinct" colors
-            # WARNING: Unique axis argument requires numpy version >=1.13
-            for name, color in loaded.items():
-                for string, replace in TRANSLATE_COLORS:
-                    if string in name:
-                        name = name.replace(string, replace)
-                if any(string in name for string in ALWAYS_REMOVE):
-                    continue  # remove "unpofessional" names
-                hcls.append(to_xyz(color, space=space))
-                filtered.append((name, color))  # category name pair
-            hcls = np.asarray(hcls)
-            if not hcls.size:
-                continue
-            hcls = hcls / np.array([360, 100, 100])
-            hcls = np.round(hcls / margin).astype(np.int64)
-            _, idxs = np.unique(hcls, return_index=True, axis=0)
-
-            # Register "distinct" colors
-            for idx in idxs:
-                name, color = filtered[idx]
-                mcolors.colorConverter.colors[name] = color
-                XKCD_COLORS[name] = color
-        else:
-            raise ValueError(f'Unknown proplot color database {path!r}.')
-
-
-def register_fonts():
-    """
-    Add fonts packaged with ProPlot or saved to the ``~/.proplot/fonts``
-    folder, if they are not already added. Detects ``.ttf`` and ``.otf`` files
-    -- see `this link \
-<https://gree2.github.io/python/2015/04/27/python-change-matplotlib-font-on-mac>`__
-    for a guide on converting various other font file types to ``.ttf`` and
-    ``.otf`` for use with matplotlib.
-
-    To visualize the registered fonts, use `~proplot.show.show_fonts`.
-    """
-    # Find proplot fonts
-    # WARNING: If you include a font file with an unrecognized style,
-    # matplotlib may use that font instead of the 'normal' one! Valid styles:
-    # 'ultralight', 'light', 'normal', 'regular', 'book', 'medium', 'roman',
-    # 'semibold', 'demibold', 'demi', 'bold', 'heavy', 'extra bold', 'black'
-    # https://matplotlib.org/api/font_manager_api.html
-    # For macOS the only fonts with 'Thin' in one of the .ttf file names
-    # are Helvetica Neue and .SF NS Display Condensed. Never try to use these!
-    import matplotlib.font_manager as mfonts
-    paths_proplot = _get_data_paths('fonts', reverse=True)
-    fnames_proplot = set(mfonts.findSystemFonts(paths_proplot))
-
-    # Detect user-input ttc fonts and issue warning
-    fnames_proplot_ttc = {
-        file for file in fnames_proplot if os.path.splitext(file)[1] == '.ttc'
-    }
-    if fnames_proplot_ttc:
-        warnings._warn_proplot(
-            'Ignoring the following .ttc fonts because they cannot be '
-            'saved into PDF or EPS files (see matplotlib issue #3135): '
-            + ', '.join(map(repr, sorted(fnames_proplot_ttc)))
-            + '. Please consider expanding them into separate .ttf files.'
-        )
-
-    # Rebuild font cache only if necessary! Can be >50% of total import time!
-    fnames_all = {font.fname for font in mfonts.fontManager.ttflist}
-    fnames_proplot -= fnames_proplot_ttc
-    if not fnames_all >= fnames_proplot:
-        warnings._warn_proplot('Rebuilding font cache.')
-        if hasattr(mfonts.fontManager, 'addfont'):
-            # New API lets us add font files manually
-            for fname in fnames_proplot:
-                mfonts.fontManager.addfont(fname)
-            mfonts.json_dump(mfonts.fontManager, mfonts._fmcache)
-        else:
-            # Old API requires us to modify TTFPATH
-            # NOTE: Previously we tried to modify TTFPATH before importing
-            # font manager with hope that it would load proplot fonts on
-            # initialization. But 99% of the time font manager just imports
-            # the FontManager from cache, so this doesn't work.
-            paths = ':'.join(paths_proplot)
-            if 'TTFPATH' not in os.environ:
-                os.environ['TTFPATH'] = paths
-            elif paths not in os.environ['TTFPATH']:
-                os.environ['TTFPATH'] += ':' + paths
-            mfonts._rebuild()
-
-    # Remove ttc files *after* rebuild
-    mfonts.fontManager.ttflist = [
-        font for font in mfonts.fontManager.ttflist
-        if os.path.splitext(font.fname)[1] != '.ttc'
-    ]
-
-    # Populate font name lists, with proplot fonts *first*
-    fonts_proplot = sorted({
-        font.name for font in mfonts.fontManager.ttflist
-        if any(path in font.fname for path in paths.split(':'))
-    })
-    fonts_system = sorted({
-        font.name for font in mfonts.fontManager.ttflist
-        if not any(path in font.fname for path in paths.split(':'))
-    })
-    fonts[:] = [*fonts_proplot, *fonts_system]
-
-# Convert colormaps that *should* be LinearSegmented from Listed
-for _name in ('viridis', 'plasma', 'inferno', 'magma', 'cividis', 'twilight'):
-    _cmap = mcm.cmap_d.get(_name, None)
-    if _cmap and isinstance(_cmap, mcolors.ListedColormap):
-        del mcm.cmap_d[_name]
-        mcm.cmap_d[_name] = LinearSegmentedColormap.from_list(
-            _name, _cmap.colors, cyclic=(_name == 'twilight')
-        )
-
-# Initialize customization folders
-_rc_folder = os.path.join(os.path.expanduser('~'), '.proplot')
-if not os.path.isdir(_rc_folder):
-    os.mkdir(_rc_folder)
-for _rc_sub in ('cmaps', 'cycles', 'colors', 'fonts'):
-    _rc_sub = os.path.join(_rc_folder, _rc_sub)
-    if not os.path.isdir(_rc_sub):
-        os.mkdir(_rc_sub)
 
 # Apply monkey patches to top level modules
 if not isinstance(mcm.cmap_d, CmapDict):
@@ -3192,24 +2933,6 @@ if not isinstance(mcolors._colors_full_map, _ColorMappingOverride):
     mcolors._colors_full_map = _map
     mcolors.colorConverter.cache = _map.cache  # re-instantiate
     mcolors.colorConverter.colors = _map  # re-instantiate
-
-#: List of registered colormap names.
-cmaps = []
-
-#: List of registered color cycle names.
-cycles = []
-
-#: Lists of registered color names by category.
-colors = {}
-
-#: Registered font names.
-fonts = []
-
-# Call driver funcs
-register_colors()
-register_cmaps()
-register_cycles()
-register_fonts()
 
 #: Dictionary of possible normalizers. See `Norm` for a table.
 normalizers = {
